refactor(cli): fold state-db import/ops helpers into state.py
Move _state_db_import.py and _state_db_ops.py function bodies directly
into state.py. Public __all__ surface and all import paths are preserved.
